Show backlog contacts only on the day they turn into backlog
Previously, backlog entries were calculated based on current date and shown in
the agenda.  This caused today's backlog entries to be shown in all entries in
the agenda, irrespective of the date in the agenda.  This commit shows all
backlog entries on the current day, and otherwise shows only the entries that
turned into backlog on the day being viewed in the agenda.

Closes #3